+++ ImmersiveCrosshair/Harmony/Adapters/HitInfoDetailsAdapter.cs
@@ -1,0 +1,14 @@
+namespace ImmersiveCrosshair.Harmony.Adapters
+{
+    public class HitInfoDetailsAdapter : IHitInfoDetails
+    {
+        private readonly HitInfoDetails _hitInfoDetails;
+
+        public HitInfoDetailsAdapter(HitInfoDetails hitInfoDetails)
+        {
+            _hitInfoDetails = hitInfoDetails;
+        }
+
+        public float distanceSq => _hitInfoDetails.distanceSq;
+    }
+}