+++ ImmersiveCrosshair/Harmony/IGuiDrawCrosshair.cs
@@ -1,0 +1,11 @@
+namespace ImmersiveCrosshair.Harmony
+{
+    public interface IGuiDrawCrosshair
+    {
+        void Update(
+            IEntityPlayerLocal entityPlayerLocal,
+            IItemAction itemAction,
+            IWorldRayHitInfo worldRayHitInfo
+        );
+    }
+}