+++ ImmersiveCrosshair/Harmony/Logger.cs
@@ -1,33 +1,40 @@
+using System;
+
 namespace ImmersiveCrosshair.Harmony
 {
     public class Logger : ILogger
     {
+        private const string LogFormat = "[ImmersiveCrosshair]: {0}";
+
         public void Info(string message)
         {
+            LogMessage(Log.Out, message);
+        }
+
+        public void Debug(string message)
+        {
 #if DEBUG
-            Log.Out("[ImmersiveCrosshair]: " + message);
+            Info(message);
 #endif
         }
 
         public void Warn(string message)
         {
 #if DEBUG
-            Log.Warning("[ImmersiveCrosshair]: " + message);
+            LogMessage(Log.Warning, message);
 #endif
         }
 
         public void Error(string message)
         {
 #if DEBUG
-            Log.Error("[ImmersiveCrosshair]: " + message);
+            LogMessage(Log.Error, message);
 #endif
         }
-    }
 
-    public interface ILogger
-    {
-        void Info(string message);
-        void Warn(string message);
-        void Error(string message);
+        private static void LogMessage(Action<string> logAction, string message)
+        {
+            logAction(string.Format(LogFormat, message));
+        }
     }
 }